requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
openai==1.51.2
requests==2.31.0
